flash-{taimen,walleye}_defconfig: Disable COMPAT_VDSO
I keep having people contacting me about the kernel build erroring with
this option enabled because CROSS_COMPILE_ARM32 needs to be set.
Lineage's kernel.mk has not been updated yet and I don't see an apparent
clean cut way to do it so I will disable it for everyone and enable it
with my script.

https://github.com/nathanchance/scripts/commit/1c7153ab9d3446fc6d772d903b5b94dbc7236e83

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>