random: Rename CANARY_MASK to RANDOM_CANARY_MASK
Qualcomm backported get_random_{u32,u64} in msm-4.9 for some kernel
hardening. WireGuard defines them for kernels older than 4.11 for
cryptography purposes so it needed to find some way to differentiate
the msm-4.9 backport from the rest of the < 4.11 kernels. CANARY_MASK
was settled on in commit a0225408e002 ("compat: work around qcom 4.9
backports"). Unfortunately, I also added CANARY_MASK as a part of some
CopperheadOS hardening (linked below).

Rename CANARY_MASK so WireGuard builds properly and we can take
advantage of the hardening offered by this. Another option would be to
backport get_random_{u32,u64} but that's not as clean as this.

https://github.com/CopperheadOS/kernel_google_wahoo/commit/0488672a4d78ca6558184e99d5dacf3c34a08dc5
https://github.com/CopperheadOS/kernel_google_wahoo/commit/d0803a5f103143cd28681a924abac7cb721e0392
https://github.com/CopperheadOS/kernel_google_wahoo/commit/b1908a4f7ea594bb2a4135970ceb4550b4269788

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>